move watch folder hosted service
